# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API circleoflinks",
        "description": "API de red social de profecionales ",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/circleoflinks-app/public"
        }
    ],
    "paths": {
...
        "/api/v1/userperfil": {
            "get": {
                "tags": [
                    "User profile"
                ],
                "summary": "Se muestran los registros de user oferta laborales del user ",
                "description": "show muestra registro especifico de postulacion oferta laboral",
                "operationId": "1f48b5b7be862fb7015a71830ef03f18",
                "responses": {
                    "200": {
                        "description": "Oferta laboral encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "info": {
                                            "type": "string",
                                            "example": "closed"
                                        },
                                        "education": {
                                            "type": "string",
                                            "example": " oferta laboral cerrada"
                                        },
                                        "exp_laboral": {
                                            "type": "string",
                                            "example": "2024-11-11"
                                        },
                                        "habilidades": {
                                            "type": "string",
                                            "example": " oferta laboral cerrada"
                                        },
                                        "profetion_name": {
                                            "type": "string",
                                            "example": "comercio exterior"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta laboral no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "status Oferta laboral no encontrada con ID: {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/userperfil/{id}": {
            "get": {
                "tags": [
                    "User profile"
                ],
                "summary": "Se muestran los registros de user oferta laborales del user ",
                "description": "show muestra registro especifico de postulacion oferta laboral",
                "operationId": "6ef48432b0e363dd0c0f3758c9407430",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Oferta laboral encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "info": {
                                            "type": "string",
                                            "example": "closed"
                                        },
                                        "education": {
                                            "type": "string",
                                            "example": " oferta laboral cerrada"
                                        },
                                        "exp_laboral": {
                                            "type": "string",
                                            "example": "2024-11-11"
                                        },
                                        "habilidades": {
                                            "type": "string",
                                            "example": " oferta laboral cerrada"
                                        },
                                        "profetion_name": {
                                            "type": "string",
                                            "example": "comercio exterior"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta laboral no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "status Oferta laboral no encontrada con ID: {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}